+++ src/main/java/org/opensearch/security/ssl/OpenSearchSecuritySSLPlugin.java
@@ -73,7 +73,7 @@ import org.opensearch.script.ScriptService;
 import org.opensearch.security.DefaultObjectMapper;
 import org.opensearch.security.NonValidatingObjectMapper;
 import org.opensearch.security.filter.SecurityRestFilter;
-import org.opensearch.security.http.SecurityHttpServerTransport;
+import org.opensearch.security.http.SecureHttpServerTransport;
 import org.opensearch.security.ssl.http.netty.ValidatingDispatcher;
 import org.opensearch.security.ssl.rest.SecuritySSLInfoAction;
 import org.opensearch.security.ssl.transport.DefaultPrincipalExtractor;
@@ -277,7 +277,7 @@ public class OpenSearchSecuritySSLPlugin extends Plugin implements SystemIndexPl
                 configPath,
                 NOOP_SSL_EXCEPTION_HANDLER
             );
-            final SecurityHttpServerTransport sgsnht = new SecurityHttpServerTransport(
+            final SecureHttpServerTransport sgsnht = new SecureHttpServerTransport(
                 migrateSettings(settings),
                 networkService,
                 bigArrays,
